endpoint for register of user
