+++ models/index.js
@@ -28,4 +28,5 @@ Object.keys(db).forEach(modelName => {
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
+
 module.exports = db;
