+++ package.json
@@ -6,12 +6,9 @@
     "start": "node ./bin/www"
   },
   "dependencies": {
-<<<<<<< HEAD
-    "bcrypt": "^5.0.1",
-=======
     "@sendgrid/mail": "^7.7.0",
+    "bcrypt": "^5.0.1",
     "bcryptjs": "^2.4.3",
->>>>>>> development
     "cookie-parser": "~1.4.4",
     "cors": "^2.8.5",
     "debug": "~2.6.9",
